+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.2.0
+
+### Added
+
+- Requires authentication headers to be set up as secrets using `PatientHttp.secret`. This is enforced and and error will be raised if plain text values are included for any of the following headers in the provider configuration: `authorization`, `x-api-key`, `x-goog-api-key`.
+
+### Changed
+
+- Requires patient_http version 1.1 or higher.
+
 ## 0.1.0
 
 ### Added